Authenticate all the things!
Everything requires authentication, but OAuth scopes haven't yet been
defined.

+++ oasis-webapp/src/main/java/oasis/web/applications/ServiceEndpoint.java
@@ -9,9 +9,11 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 
 import com.google.common.base.Strings;
 import com.wordnik.swagger.annotations.Api;
@@ -20,18 +22,26 @@ import com.wordnik.swagger.annotations.ApiOperation;
 import oasis.model.InvalidVersionException;
 import oasis.model.applications.v2.Service;
 import oasis.model.applications.v2.ServiceRepository;
+import oasis.model.directory.OrganizationMembership;
+import oasis.model.directory.OrganizationMembershipRepository;
 import oasis.services.etag.EtagService;
 import oasis.web.authn.Authenticated;
 import oasis.web.authn.OAuth;
+import oasis.web.authn.OAuthAuthenticationFilter;
+import oasis.web.authn.OAuthPrincipal;
 import oasis.web.utils.ResponseFactory;
 
 @Path("/apps/service/{service_id}")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
+@OAuth
 @Api(value = "services", description = "Application services")
 public class ServiceEndpoint {
   @Inject ServiceRepository serviceRepository;
+  @Inject OrganizationMembershipRepository organizationMembershipRepository;
   @Inject EtagService etagService;
+
+  @Context SecurityContext securityContext;
 
   @PathParam("service_id") String serviceId;
 
@@ -44,6 +54,20 @@ public class ServiceEndpoint {
     Service service = serviceRepository.getService(serviceId);
     if (service == null) {
       return ResponseFactory.notFound("Service not found");
+    }
+
+    if (!service.isVisible()) {
+      // only the admins or designed users should be able to see the service if it's "hidden"
+      // TODO: check app_users; for now app_users are all members of the organization (and app_admins ⊆ app_users)
+      OAuthPrincipal principal = (OAuthPrincipal) securityContext.getUserPrincipal();
+      if (principal == null) {
+        return OAuthAuthenticationFilter.challengeResponse();
+      }
+      String userId = principal.getAccessToken().getAccountId();
+      OrganizationMembership organizationMembership = organizationMembershipRepository.getOrganizationMembership(userId, service.getProvider_id());
+      if (organizationMembership == null) {
+        return ResponseFactory.forbidden("Current user is neither an app_admin or app_user for the service");
+      }
     }
 
     // XXX: don't send secrets over the wire
@@ -60,7 +84,7 @@ public class ServiceEndpoint {
       value = "Updates the service",
       response = Service.class
   )
-  @Authenticated @OAuth
+  @Authenticated
   public Response update(
       @HeaderParam(HttpHeaders.IF_MATCH) String ifMatch,
       Service service
@@ -75,7 +99,12 @@ public class ServiceEndpoint {
     if (updatedService == null) {
       return ResponseFactory.NOT_FOUND;
     }
-    // TODO: check the user is an admin of the service's providing organization (TODO: support services bought by individuals)
+
+    // TODO: support applications bought by individuals
+    if (!isAdminOfOrganization(((OAuthPrincipal) securityContext.getUserPrincipal()).getAccessToken().getAccountId(), updatedService.getProvider_id())) {
+      return ResponseFactory.forbidden("Current user is not an admin of the service's providing organization");
+    }
+
     try {
       service = serviceRepository.updateService(service, etagService.parseEtag(ifMatch));
     } catch (InvalidVersionException e) {
@@ -93,13 +122,23 @@ public class ServiceEndpoint {
 
   @DELETE
   @ApiOperation("Deletes the service")
+  @Authenticated
   public Response delete(
       @HeaderParam(HttpHeaders.IF_MATCH) String ifMatch
   ) {
     if (Strings.isNullOrEmpty(ifMatch)) {
       return ResponseFactory.preconditionRequiredIfMatch();
     }
-    // TODO: check the user is an admin of the service's providing organization (TODO: support services bought by individuals)
+
+    // TODO: support applications bought by individuals
+    Service serviceToBeDeleted = serviceRepository.getService(serviceId);
+    if (serviceToBeDeleted == null) {
+      return ResponseFactory.NOT_FOUND;
+    }
+    if (!isAdminOfOrganization(((OAuthPrincipal) securityContext.getUserPrincipal()).getAccessToken().getAccountId(), serviceToBeDeleted.getProvider_id())) {
+      return ResponseFactory.forbidden("Current user is not an admin of the service's providing organization");
+    }
+
     boolean deleted;
     try {
       deleted = serviceRepository.deleteService(serviceId, etagService.parseEtag(ifMatch));
@@ -110,5 +149,13 @@ public class ServiceEndpoint {
       return ResponseFactory.NOT_FOUND;
     }
     return ResponseFactory.NO_CONTENT;
+  }
+
+  private boolean isAdminOfOrganization(String userId, String organizationId) {
+    OrganizationMembership organizationMembership = organizationMembershipRepository.getOrganizationMembership(userId, organizationId);
+    if (organizationMembership == null) {
+      return false;
+    }
+    return organizationMembership.isAdmin();
   }
 }
